쪽지보관함

+++ __init__.py
@@ -1,14 +1,14 @@
-from flask import Flask,render_template, request, redirect, url_for, session,make_response,jsonify
+from flask import Flask,render_template, request, redirect, url_for, session,make_response,jsonify,flash
 from werkzeug.security import generate_password_hash, check_password_hash
 from model import Customer_DAO
 from predict import CelebrityPredictionModel
-from datetime import datetime
+from datetime import datetime,timedelta
 app = Flask(__name__)
 
-
+app.config["PERMANENT_SESSION_LIFETIME"] = timedelta(minutes=10) 
 predictor = CelebrityPredictionModel("./model")
 
-# 회우너가입에서 ID
+# 회원가입 ID 체크
 @app.route('/register/check/id', methods=['POST'])
 def check_register_userid():
     if request.method=='POST':
@@ -31,6 +31,50 @@ def check_register_nickname():
                 return make_response(jsonify(result,201))
     else:
         return redirect('/',200)
+
+@app.route('/register',methods=['GET','POST'])
+def register():
+    if 'userid' not in session:
+        print(request.form)
+        if request.method == 'POST':
+            id = request.form['user_id']
+            if id.isalnum() == False:
+                result = {
+                    'space' : True,
+                    'msg ' : '공백 문자 존재'
+                }
+                return make_response(jsonify(result),401)
+            if Customer_DAO.duplicate_member(request.form['userid']):
+                password =  generate_password_hash(request.form['password'])
+                if Customer_DAO.register_member(request.form['userid'],password,request.form['user_nicname'],request.form['user_fullname'],request.form['birthday'],request.form['paper_flag'],request.form['gender_flag'],request.form['create_at']) :
+                    return '성공'
+                    # return redirect('/',200)
+                else:
+                    return '실패'
+        return render_template('register.html',msg='register loading succ')
+
+@app.route('/login',methods=['GET','POST'])
+def login():
+
+    if 'user_id' in session:
+        return render_template(url_for('/'))
+    if request.method == 'POST':
+        print(request.method)
+        print(request.form['userid'])
+    
+        password = request.form['password']
+        if Customer_DAO.check_login(request.form['userid'],password):
+            print(request.form)
+            session['user_id'] = request.form['userid']
+            res = make_response(redirect('/'))
+            res.set_cookie('userid',request.form['userid'])
+            return redirect('/')
+        else:
+            flash("아이디와 비밀번호를 확인해 주세요")
+            return redirect(url_for('login'))
+    return render_template('loginM.html',msg='testing now')
+
+
 @app.route('/',methods=['GET','POST'])
 def main_view():
     # 
@@ -42,7 +86,10 @@ def main_view():
             return render_template('main.html',id= '')
 
     elif request.method =='GET':
-        return render_template('main.html')
+        if 'user_id' in session:
+            return render_template('main.html',id=session['user_id'])
+        else:
+            return render_template('main.html')
 # @app.route('/mypage/update', methods=['POST'])
 #     # if session['user_id']:
 #     #     if request.
@@ -62,48 +109,20 @@ def mypage():
     else:
         #  세션이 존재 x이면 닮은꼴 입력하는 초기 화면으로 진행한다.
         return redirect('/')
-@app.route('/register',methods=['GET','POST'])
-def register():
-    print(request.form)
-    if request.method == 'POST':
-        id = request.form['user_id']
-        if id.isalnum() == False:
-            result = {
-                'space' : True,
-                'msg ' : '공백 문자 존재'
-            }
-            return make_response(jsonify(result),401)
-        if Customer_DAO.duplicate_member(request.form['userid']):
-            password =  generate_password_hash(request.form['password'])
-            if Customer_DAO.register_member(request.form['userid'],password,request.form['user_nicname'],request.form['user_fullname'],request.form['birthday'],request.form['paper_flag'],request.form['gender_flag'],request.form['create_at']) :
-                return '성공'
-                # return redirect('/',200)
-            else:
-                return '실패'
-    return render_template('register.html',msg='register loading succ')
+@app.route('/person',methods=['GET'])
+def person():
+    return render_template('pri.html')
 
-@app.route('/login',methods=['GET','POST'])
-def login():
-    print(request.form)
-    if 'id' in session:
-        return render_template('main.html')
-    if request.method == 'POST':
-        password = request.form['password']
-        if Customer_DAO.check_login(request.form['username'],password):
-            print(request.form)
-            session['user_id'] = request.form['username']
-            return '로그인 성공'
-        else:
-            return '로그인 실패'
-    return render_template('login.html',msg='testing now')
 
 #  get all은 쪽지함 조회 클릭시 기본적으로 보여주는 곳
 @app.route('/note/all',methods=['GET','POST'])
 def notes():
     if session['user_id']:
         return redirect('/',406)
-
-
+@app.route('/logout',methods=['GET','POST'])
+def logout():
+    session.clear()
+    return redirect('/')
 # @app.route('/notice/share',methods=['GET','POST'])
 # def share_notice():
 #     pass
@@ -132,6 +151,10 @@ def temp():
             # save_to = f'static/img/profiles/{user_id}'
             f.save(save_to)
             result = predictor.predict_img(gender,save_to)
+            if result == []:
+                print("flash")
+                flash('Not find humon')
+                return render_template('temp.html')
             if os.path.isfile(save_to):
                 os.remove(save_to)
                 print('file delete finish')
@@ -141,4 +164,6 @@ def temp():
         return render_template('temp.html')
 
 if __name__== '__main__':
+    app.secret_key = 'super secret key'
+    app.config['SESSION_TYPE'] = 'filesystem'
     app.run(debug = True, port=8080)
